bug(fix heroku build):fix bug in heroku build - Implement a new script that builds webpack in server side. - Implement a new script that builds webpack in client side. - Implement rebuild inside dist folder. [Fixes #160819372]
