+++ client/actions/CreateRideAction.js
@@ -1,7 +1,8 @@
 import axios from 'axios';
+import config from '../config';
 
 export const createRide = (rideData) => {
     return dispatch => {
-        return axios.post('https://frozen-mesa-95948.herokuapp.com/api/v1/users/rides', rideData);
+        return axios.post(`${config.apiUrl}/api/v1/users/rides`, rideData);
     };
 };
