+++ client/actions/FetchRidesAction.js
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import config from '../config';
 
 export const ALL_RIDES = 'ALL_RIDES';
 
@@ -12,7 +13,7 @@ const allRides = ( rides ) => {
 
 const fetchRides = () => {
     return dispatch => {
-        return axios.get('http://localhost:3000/api/v1/rides').then( response => {
+        return axios.get(`${config.apiUrl}/api/v1/rides`).then( response => {
             dispatch(allRides(response.data.rides));
         }).catch( error => {
             throw(error);
