+++ client/actions/JoinRideAction.js
@@ -1,7 +1,8 @@
 import axios from 'axios';
+import config from '../config';
 
-export const joinRide = (joinRideData) => {
-  return dispatch => {
-    return axios.post('https://frozen-mesa-95948.herokuapp.com/api/v1//rides/:rideId/requests')
-  }
-}
+export const joinRide = (rideId) => {
+    return dispatch => {
+        return axios.post(`${config.apiUrl}/api/v1//rides/${rideId}/requests`);
+    };
+};
