+++ client/actions/LoginAction.js
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import jwt from 'jsonwebtoken';
+import config from '../config';
 import { SET_CURRENT_USER } from './types.js';
 import setAuthorizationToken from '../utils/setAuthorizationToken.js';
 
@@ -20,7 +21,7 @@ export function logout() {
 
 export function userLoginRequest(userLoginData) {
     return (dispatch) => {
-        return axios.post('http://localhost:3000/api/v1/auth/login', userLoginData).then(
+        return axios.post(`${config.apiUrl}/api/v1/auth/login`, userLoginData).then(
             res => {
                 const token = res.data.data;
                 localStorage.setItem('jwtToken', token);
