+++ client/actions/signUpAction.js
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import config from '../config';
 import jwt from 'jsonwebtoken';
 import { SET_CURRENT_USER } from './types.js';
 import setAuthorizationToken from '../utils/setAuthorizationToken.js';
@@ -12,7 +13,7 @@ export function setCurrentUser(user){
 
 export function userSignUpRequest(userData) {
     return (dispatch) => {
-        return axios.post('https://frozen-mesa-95948.herokuapp.com/api/v1/auth/signup', userData).then(
+        return axios.post(`${config.apiUrl}/api/v1/auth/signup`, userData).then(
             res => {
                 const token = res.data.token;
                 localStorage.setItem('jwtToken', token);
