+++ client/config/index.js
@@ -1,0 +1,3 @@
+export default {
+    apiUrl: process.env.NODE_ENV === 'production' ? '': 'http://localhost:3000',
+};
