+++ client/index.html
@@ -11,7 +11,5 @@
 </head>
 <body>
     <div id="app"></div>
-
-    <script type ="text/javascript" src="/bundle.min.js"></script>
 </body>
 </html>
