+++ package.json
@@ -6,16 +6,18 @@
   "scripts": {
     "postinstall": "npm run build:prod",
     "start:dev": "nodemon ./server/app.js --exec babel-node",
-    "start": "concurrently \"node dist/app.js\" \"npm run dev\"",
+    "start": "node dist/app.js",
     "clean:dist": "rimraf ./dist",
-    "build:prod": "npm run clean:dist && babel ./server --out-dir dist",
+    "build:prod": "npm run clean:dist && babel ./server --out-dir dist --copy-files",
     "test": " nyc --reporter=html --reporter=text mocha --compilers js:babel-core/register server/test/test/index.test.js --timeout 10000ms",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "coveralls": "nyc --reporter=lcov --reporter=text-lcov npm test",
     "createDb": "psql -U postgres -f ./server/db/data.sql",
-    "dev": "./node_modules/.bin/webpack-dev-server --config ./webpack.config.dev.js",
+    "dev": "./node_modules/.bin/webpack-dev-server --config ./webpack.config.js",
     "start:concurrent": "concurrently \"npm run start:dev\" \"npm run dev\"",
-    "heroku-postbuild": "npm run dev"
+    "build:client": "webpack --mode=production",
+    "build": "npm run build:client && npm run build:prod",
+    "heroku-postbuild": "npm run build"
   },
   "engines": {
     "node": "10.4.1"
@@ -83,6 +85,8 @@
     "eslint-config-airbnb": "^15.1.0",
     "eslint-plugin-import": "^2.7.0",
     "eslint-plugin-react": "^7.1.0",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.2.0",
     "istanbul": "^0.4.5",
     "mocha-istanbul": "^0.3.0",
     "mocha-lcov-reporter": "^1.3.0",
